+++ src/calmjs/tests/test_runtime.py
@@ -58,6 +58,29 @@ class DeprecatedRuntime(runtime.DriverRuntime):
 deprecated = DeprecatedRuntime(None)
 
 
+class DefaultCommandRuntime(runtime.Runtime):
+
+    def __init__(
+            self, entry_point_group='broken.test', action_key='broken_test',
+            *a, **kw):
+        super(DefaultCommandRuntime, self).__init__(
+            entry_point_group=entry_point_group, action_key=action_key,
+            *a, **kw)
+
+    def init_argparser(self, argparser):
+        argparser.add_argument(
+            '--valid', action='store_true', required=False)
+        return super(DefaultCommandRuntime, self).init_argparser(argparser)
+
+
+dummy = runtime.DriverRuntime(None)
+default_cmd = DefaultCommandRuntime(working_set=mocks.WorkingSet({
+    'broken.test': [
+        'dummy = calmjs.tests.test_runtime:dummy',
+    ],
+}))
+
+
 class BaseRuntimeTestCase(unittest.TestCase):
 
     def tearDown(self):
@@ -149,6 +172,18 @@ class BaseRuntimeTestCase(unittest.TestCase):
         rt = runtime.Runtime(working_set=working_set, prog='dummy')
         result = rt.run(rt.argparser, runtime='dummy')
         self.assertIs(result, NotImplemented)
+
+    def test_runtime_error_deprecation(self):
+        stub_stdouts(self)
+        working_set = mocks.WorkingSet({'calmjs.runtime': [
+            'deprecated = calmjs.tests.test_runtime:deprecated',
+        ]})
+        rt = runtime.Runtime(working_set=working_set, prog='dummy')
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter('always')
+            with self.assertRaises(SystemExit):
+                rt.error(rt.argparser, 'deprecated', 'simulated')
+        self.assertIn('Runtime.error is deprecated', str(w[-1].message))
 
     def test_runtime_run_abort(self):
         class CustomRuntime(runtime.BaseRuntime):
@@ -905,7 +940,8 @@ class ArtifactRuntimeTestCase(unittest.TestCase):
             'artifact = calmjs.runtime:artifact',
         ]})
         rt = runtime.Runtime(working_set=working_set)
-        rt(['artifact'])
+        # An underspecified command should also return False.
+        self.assertFalse(rt(['artifact']))
         # ensure the help for the command itself is printed
         self.assertIn(
             'helpers for the management of artifacts', sys.stdout.getvalue())
@@ -1126,7 +1162,7 @@ class PackageManagerRuntimeTestCase(unittest.TestCase):
         working_set = mocks.WorkingSet({'calmjs.runtime': [
             'npm = calmjs.npm:npm.runtime',
         ]})
-        rt = runtime.Runtime(working_set=working_set)
+        rt = runtime.Runtime(prog='dummy', working_set=working_set)
         rt.argparser  # populate the argparser
         rt.argparser_details.clear()
         with pretty_logging(
@@ -1135,7 +1171,8 @@ class PackageManagerRuntimeTestCase(unittest.TestCase):
 
         self.assertIn('CRITICAL', s.getvalue())
         self.assertIn(
-            'provided argparser not registered to this runtime.',
+            "provided argparser (prog='dummy') not associated with this "
+            "runtime (<calmjs.runtime.Runtime",
             s.getvalue())
         self.assertIn(
             'runtime cannot continue due to missing argparser details',
@@ -1645,7 +1682,7 @@ class RuntimeGoingWrongTestCase(unittest.TestCase):
         self.assertIn("CRITICAL", stderr)
         self.assertIn(
             "Runtime subclass at entry_point 'badsimple = "
-            "calmjs.testing.utils:badsimple' has overide "
+            "calmjs.testing.utils:badsimple' has override "
             "'entry_point_load_validated' without filtering out "
             "its parent classes; this can be addressed by calling "
             "super(calmjs.testing.module3.runtime.BadSimpleRuntime, self)."
@@ -1826,10 +1863,11 @@ class ArgumentHandlingTestCase(unittest.TestCase):
     def test_before_and_after_extras(self):
         rt = self.setup_runtime()
         with self.assertRaises(SystemExit):
-            rt(['-u', 'cmd', 'pkg', '-u'])
-        err = sys.stderr.getvalue().splitlines()[-1].strip()
-        # former has priority
-        self.assertEqual("calmjs: error: unrecognized arguments: -u", err)
+            rt(['--view', 'cmd', 'pkg', '-n', '-v'])
+        err = sys.stderr.getvalue()
+        # --view is recognized in calmjs cmd
+        self.assertIn("calmjs: error: unrecognized arguments: --view", err)
+        self.assertIn("calmjs cmd: error: unrecognized arguments: -n", err)
 
     def test_before_known_and_after_unknown(self):
         rt = self.setup_runtime()
@@ -1837,6 +1875,16 @@ class ArgumentHandlingTestCase(unittest.TestCase):
             rt(['-v', 'cmd', 'pkg', '-u'])
         err = sys.stderr.getvalue().splitlines()[-1].strip()
         self.assertEqual("calmjs cmd: error: unrecognized arguments: -u", err)
+
+    def test_before_known_to_after_but_not_after(self):
+        rt = self.setup_runtime()
+        with self.assertRaises(SystemExit):
+            rt(['--view', 'cmd', 'pkg'])
+        err = sys.stderr.getvalue()
+        errline = err.splitlines()[-1].strip()
+        self.assertNotIn('calmjs cmd: error:', err)
+        self.assertEqual(
+            "calmjs: error: unrecognized arguments: --view", errline)
 
     def test_before_and_after_extras_scattered(self):
         # previously test_before_and_after_extras_known_before, as
@@ -1865,6 +1913,121 @@ class ArgumentHandlingTestCase(unittest.TestCase):
         err = sys.stderr.getvalue().splitlines()[-1].strip()
         # exact message differs between py2 and py3
         self.assertIn("calmjs cmd: error: ", err)
+
+    def test_subparser_level_2_missing_argument(self):
+        stub_stdouts(self)
+        working_set = mocks.WorkingSet({
+            'calmjs.runtime': [
+                'artifact = calmjs.runtime:artifact',
+            ],
+            'calmjs.runtime.artifact': [
+                'build = calmjs.runtime:artifact_build',
+            ],
+        })
+        rt = runtime.Runtime(working_set=working_set, prog='calmjs')
+        # An underspecified command should also return False.
+        with self.assertRaises(SystemExit):
+            rt(['artifact', 'build'])
+        # ensure the help for the command itself is printed
+        self.assertIn('calmjs artifact build: error', sys.stderr.getvalue())
+
+    def test_subparser_level_2_unrecognized_argument_final(self):
+        stub_stdouts(self)
+        working_set = mocks.WorkingSet({
+            'calmjs.runtime': [
+                'artifact = calmjs.runtime:artifact',
+            ],
+            'calmjs.runtime.artifact': [
+                'build = calmjs.runtime:artifact_build',
+            ],
+        })
+        rt = runtime.Runtime(working_set=working_set, prog='calmjs')
+        # An underspecified command should also return False.
+        with self.assertRaises(SystemExit):
+            rt(['artifact', 'build', 'package', '--no-such-argument'])
+        # ensure the help for the command itself is printed
+        self.assertIn('calmjs artifact build: error', sys.stderr.getvalue())
+
+    def test_subparsers_unrecognized_argument_interspersed(self):
+        working_set = mocks.WorkingSet({
+            'calmjs.runtime': [
+                'broken = calmjs.tests.test_runtime:default_cmd',
+            ],
+        })
+        rt = runtime.Runtime(working_set=working_set, prog='calmjs')
+        # An underspecified command should also return False.
+        with self.assertRaises(SystemExit):
+            rt(['--1', 'broken', '--2', 'dummy', '--3'])
+        # ensure the help for the command itself is printed
+        err = sys.stderr.getvalue()
+        self.assertIn(
+            'calmjs: error: unrecognized arguments: --1', err)
+        self.assertIn(
+            'calmjs broken: error: unrecognized arguments: --2', err)
+        self.assertIn(
+            'calmjs broken dummy: error: unrecognized arguments: --3', err)
+
+    def test_subparsers_unrecognized_argument_skipped(self):
+        rt = runtime.Runtime(prog='calmjs')
+        # An underspecified command should also return False.
+        with self.assertRaises(SystemExit):
+            rt(['--1', 'artifact', 'build', 'package', '--3'])
+        # ensure the help for the command itself is printed
+        err = sys.stderr.getvalue()
+        self.assertIn(
+            'calmjs: error: unrecognized arguments: --1', err)
+        self.assertNotIn('calmjs artifact: error', err)
+        self.assertIn(
+            'calmjs artifact build: error: unrecognized arguments: --3', err)
+
+    def test_subparsers_first_fail(self):
+        rt = runtime.Runtime(prog='calmjs')
+        # An underspecified command should also return False.
+        with self.assertRaises(SystemExit):
+            rt(['--1', 'artifact', 'build', 'package'])
+        # ensure the help for the command itself is printed
+        err = sys.stderr.getvalue()
+        self.assertIn('calmjs: error: unrecognized arguments: --1', err)
+        self.assertNotIn('calmjs artifact: error', err)
+        self.assertNotIn('calmjs artifact build: error', err)
+
+    def test_subparsers_first_also_missing(self):
+        rt = runtime.Runtime(prog='calmjs')
+        # An underspecified command should also return False.
+        with self.assertRaises(SystemExit):
+            rt(['--1', 'artifact', '-v', '--2'])
+        # ensure the help for the command itself is printed
+        err = sys.stderr.getvalue()
+        self.assertIn('calmjs: error: unrecognized arguments: --1', err)
+        self.assertIn(
+            'calmjs artifact: error: unrecognized arguments: --2', err)
+
+    def test_subparsers_unrecognized_argument_issue(self):
+        working_set = mocks.WorkingSet({
+            'calmjs.runtime': [
+                'broken = calmjs.tests.test_runtime:default_cmd',
+            ],
+        })
+        rt = runtime.Runtime(working_set=working_set, prog='calmjs')
+        # Emulating a "default" choice for a subparser, however it seems
+        # like the default parsing behavior is to do everything on the
+        # root parser until a some bare string is encountered which will
+        # immediately be assigned to the command.  Still going to spit
+        # this warning out, just in case as this is really not a typical
+        # situation.
+        rt.argparser._subparsers._actions[-1].default = 'broken'
+        with self.assertRaises(SystemExit):
+            rt(['-vv', '--2', '--3', '--valid'])
+        # ensure the help for the command itself is printed
+        err = sys.stderr.getvalue()
+        self.assertIn(
+            "DEBUG calmjs.runtime command for prog='calmjs' is set to "
+            "'broken' without being specified as part of the input arguments "
+            "- the following error message may contain misleading references",
+            err
+        )
+        self.assertIn(
+            'calmjs: error: unrecognized arguments: --2 --3 --valid', err)
 
 
 class PackageManagerRuntimeAlternativeIntegrationTestCase(unittest.TestCase):
@@ -2361,7 +2524,7 @@ class MainIntegrationTestCase(unittest.TestCase):
         with self.assertRaises(SystemExit) as e:
             runtime.main([])
         self.assertIn('usage', sys.stdout.getvalue())
-        self.assertEqual(e.exception.args[0], 0)
+        self.assertEqual(e.exception.args[0], 1)
 
     def test_calmjs_main_console_entry_point_help(self):
         stub_stdouts(self)